+++ src/QuietWheel.js
@@ -8,45 +8,51 @@ var QuietWheel = function(callback, callbackArgs, enableScrollEvents){
 
 	enableScrollEvents = enableScrollEvents || false;
 
-	var wheelUnits = [];
+	var qweUnits = [];
 	var isWheelFired = false;
 	var prevTime = new Date().getTime();
 
 	function onWheel(e){
-		if (!enableScrollEvents) e.preventDefault();
-		var curTime = new Date().getTime();
-		var wheelUnit = new WheelUnit(e);
 
-		if(wheelUnits.length > 149){
-			wheelUnits.shift();
+		if (typeof enableScrollEvents === 'function') {
+			if (!enableScrollEvents()) e.preventDefault();
+		} else {
+			if (!enableScrollEvents) e.preventDefault();
 		}
-		wheelUnits.push(wheelUnit);
+
+		var curTime = new Date().getTime();
+		var qweUnit = new QWEventUnit(e);
+
+		if(qweUnits.length > 149){
+			qweUnits.shift();
+		}
+		qweUnits.push(qweUnit);
 
 		var timeDiff = curTime-prevTime;
 		prevTime = curTime;
 
 		if(timeDiff > 200){
-			wheelUnits = [];
+			qweUnits = [];
 			isWheelFired = false;
 		}
 
-		var averageEnd = getAverageScala(wheelUnits, 10);
-		var averageMiddle = getAverageScala(wheelUnits, 70);
+		var averageEnd = getAverageScala(qweUnits, 10);
+		var averageMiddle = getAverageScala(qweUnits, 70);
 		var isAccelerating = averageEnd >= averageMiddle;
 
-		if (isAccelerating && wheelUnit.isVerticalScroll){
+		if (isAccelerating && qweUnit.isVerticalScroll){
 			if (!isWheelFired) {
 				isWheelFired = true;
-				callback.apply(e, callbackArgs);
+				callback.apply(null, [qweUnit].concat(callbackArgs));
 			} 
 		} else {
 			isWheelFired = false;
 		}
 	}
 
-	function getAverageScala(wheelUnits, range){
+	function getAverageScala(qweUnits, range){
 		var sum = 0;
-		var arr_slice = wheelUnits.slice(Math.max(wheelUnits.length - range, 1));
+		var arr_slice = qweUnits.slice(Math.max(qweUnits.length - range, 1));
 		for(var i = 0; i < arr_slice.length; i++){
 			sum += arr_slice[i].scala;
 		}
@@ -56,15 +62,15 @@ var QuietWheel = function(callback, callbackArgs, enableScrollEvents){
 	window.addEventListener('wheel', onWheel);
 };
 
-function WheelUnit(e){
+function QWEventUnit(e){
 
 	var self = this;
 
-	self.e = e;
+	self.originalWheelEvent = e;
 
 	self.scala = Math.abs(e.deltaY);
 
-	self.direction = Math.max(-1, Math.min(1, self.scala)) > 0 ? 'up' : 'down';
+	self.direction = Math.max(-1, Math.min(1, e.deltaY)) < 0 ? 'up' : 'down';
 
 	self.isVerticalScroll = Math.abs(e.deltaY) > Math.abs(e.deltaX);
 }