+++ CHANGELOG.md
@@ -1,7 +1,15 @@
 CHANGELOG
 =========
 
-## 0.1.1 (unreleased)
+## 0.1.2 (unreleased)
+
+## 0.1.1 (2017-03-02, beta release)
+
+#### features:
+ - enableScrollEvents parameter can accept either boolean or function. If function, it applies the return value as the parameter.
+
+#### updates:
+ - QWEventUnit(e) contains originalWheelEvent, scala, direction, and isVerticalScroll
 
 ## 0.1.0 (2017-03-01, beta release)